parser terminado. Valgrind OK, norminette OK
